+++ CHANGES.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support Ruby-3.4.X
+
+
 ## [0.5.2] - 2024-08-30
 
 ### Changed