+++ CHANGES.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.6.0] - 2024-12-15
+
 ### Added
 
 - Support Ruby-3.4.X.
@@ -82,6 +85,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
+[Unreleased]: https://github.com/ixti/amazing-activist/compare/v0.6.0...main
+[0.6.0]: https://github.com/ixti/amazing-activist/compare/v0.5.2...v0.6.0
+[0.5.2]: https://github.com/ixti/amazing-activist/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/ixti/amazing-activist/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/ixti/amazing-activist/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/ixti/amazing-activist/compare/v0.3.0...v0.4.0